+++ android/engine/src/main/java/org/smartregister/fhircore/engine/configuration/view/ButtonProperties.kt
@@ -47,11 +47,11 @@ data class ButtonProperties(
   val enabled: String = "true",
   val text: String? = null,
   val status: String,
-  val smallSized: Boolean = false,
   val fontSize: Float = 14.0f,
   val actions: List<ActionConfig> = emptyList(),
   val buttonType: ButtonType = ButtonType.MEDIUM,
   val startIcon: ImageConfig? = null,
+  val letterSpacing: Int = 0,
 ) : ViewProperties(), Parcelable {
   /**
    * This function determines the status color to display depending on the value of the service