+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/register/RegisterFragment.kt
@@ -32,6 +32,7 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
+import androidx.compose.ui.platform.testTag
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
@@ -50,6 +51,7 @@ import kotlinx.coroutines.flow.emptyFlow
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
+import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.smartregister.fhircore.engine.domain.model.SnackBarMessageConfig
 import org.smartregister.fhircore.engine.sync.OnSyncListener
 import org.smartregister.fhircore.engine.sync.SyncListenerManager
@@ -160,7 +162,7 @@ class RegisterFragment : Fragment(), OnSyncListener {
               )
             },
           ) { innerPadding ->
-            Box(modifier = Modifier.padding(innerPadding)) {
+            Box(modifier = Modifier.padding(innerPadding).testTag(REGISTER_SCREEN_BOX_TAG)) {
               RegisterScreen(
                 openDrawer = openDrawer,
                 onEvent = registerViewModel::onEvent,
@@ -244,11 +246,15 @@ class RegisterFragment : Fragment(), OnSyncListener {
     }
   }
 
-  fun refreshRegisterData() {
+  fun refreshRegisterData(questionnaireResponse: QuestionnaireResponse? = null) {
     with(registerFragmentArgs) {
       registerViewModel.run {
-        // Clear pages cache to load new data
+        if (questionnaireResponse != null) {
+          updateRegisterFilterState(registerId, questionnaireResponse)
+        }
+
         pagesDataCache.clear()
+
         retrieveRegisterUiState(
           registerId = registerId,
           screenTitle = screenTitle,
@@ -277,20 +283,24 @@ class RegisterFragment : Fragment(), OnSyncListener {
   }
 
   suspend fun handleQuestionnaireSubmission(questionnaireSubmission: QuestionnaireSubmission) {
-    appMainViewModel.run {
-      onQuestionnaireSubmission(questionnaireSubmission)
-      retrieveAppMainUiState(refreshAll = false) // Update register counts
+    if (questionnaireSubmission.questionnaireConfig.saveQuestionnaireResponse) {
+      appMainViewModel.run {
+        onQuestionnaireSubmission(questionnaireSubmission)
+        retrieveAppMainUiState(refreshAll = false) // Update register counts
+      }
+
+      val (questionnaireConfig, _) = questionnaireSubmission
+
+      refreshRegisterData()
+
+      questionnaireConfig.snackBarMessage?.let { snackBarMessageConfig ->
+        registerViewModel.emitSnackBarState(snackBarMessageConfig)
+      }
+
+      questionnaireConfig.onSubmitActions?.handleClickEvent(navController = findNavController())
+    } else {
+      refreshRegisterData(questionnaireSubmission.questionnaireResponse)
     }
-
-    val (questionnaireConfig, _) = questionnaireSubmission
-
-    refreshRegisterData()
-
-    questionnaireConfig.snackBarMessage?.let { snackBarMessageConfig ->
-      registerViewModel.emitSnackBarState(snackBarMessageConfig)
-    }
-
-    questionnaireConfig.onSubmitActions?.handleClickEvent(navController = findNavController())
   }
 
   fun emitPercentageProgress(
@@ -316,8 +326,8 @@ class RegisterFragment : Fragment(), OnSyncListener {
         1L,
       )
     val isProgressTotalLess = progressSyncJobStatus.total <= totalRecordsOverall
-    var currentProgress: Int
-    var currentTotalRecords =
+    val currentProgress: Int
+    val currentTotalRecords =
       if (isProgressTotalLess) {
         currentProgress =
           totalRecordsOverall.toInt() - progressSyncJobStatus.total +
@@ -334,5 +344,9 @@ class RegisterFragment : Fragment(), OnSyncListener {
       }
 
     return getSyncProgress(currentProgress, currentTotalRecords)
+  }
+
+  companion object {
+    const val REGISTER_SCREEN_BOX_TAG = "fragmentRegisterScreenTestTag"
   }
 }