+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/report/measure/screens/MeasureReportResultScreen.kt
@@ -135,7 +135,7 @@ fun MeasureReportResultPage(
         if (measureReportPopulationResult != null) {
           println("FIKRI 2")
           MeasureReportPopulationResultView(
-            measureReportPopulationResult.distinctBy { it.indicatorTitle },
+            measureReportPopulationResult.distinctBy { it.title },
           )
           println("FIKRI 2.1")
         }