+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/questionnaire/QuestionnaireViewModelTest.kt
@@ -21,6 +21,7 @@ import androidx.test.core.app.ApplicationProvider
 import com.google.android.fhir.FhirEngine
 import com.google.android.fhir.datacapture.mapping.ResourceMapper
 import com.google.android.fhir.db.ResourceNotFoundException
+import com.google.android.fhir.get
 import com.google.android.fhir.logicalId
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
@@ -40,13 +41,17 @@ import java.util.UUID
 import javax.inject.Inject
 import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.test.runTest
 import org.hl7.fhir.r4.model.Address
+import org.hl7.fhir.r4.model.Basic
 import org.hl7.fhir.r4.model.Bundle
 import org.hl7.fhir.r4.model.CodeableConcept
 import org.hl7.fhir.r4.model.Coding
+import org.hl7.fhir.r4.model.Encounter
 import org.hl7.fhir.r4.model.Enumerations
 import org.hl7.fhir.r4.model.Extension
+import org.hl7.fhir.r4.model.Flag
 import org.hl7.fhir.r4.model.Group
 import org.hl7.fhir.r4.model.IdType
 import org.hl7.fhir.r4.model.IntegerType
@@ -78,6 +83,7 @@ import org.smartregister.fhircore.engine.rulesengine.ResourceDataRulesExecutor
 import org.smartregister.fhircore.engine.task.FhirCarePlanGenerator
 import org.smartregister.fhircore.engine.util.SharedPreferenceKey
 import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
+import org.smartregister.fhircore.engine.util.extension.appendPractitionerInfo
 import org.smartregister.fhircore.engine.util.extension.asReference
 import org.smartregister.fhircore.engine.util.extension.decodeResourceFromString
 import org.smartregister.fhircore.engine.util.extension.extractLogicalIdUuid
@@ -170,7 +176,7 @@ class QuestionnaireViewModelTest : RobolectricTest() {
           transformSupportServices = mockk(),
           dispatcherProvider = defaultRepository.dispatcherProvider,
           sharedPreferencesHelper = sharedPreferencesHelper,
-          libraryEvaluator = libraryEvaluator,
+          libraryEvaluatorProvider = { libraryEvaluator },
           fhirCarePlanGenerator = fhirCarePlanGenerator,
           resourceDataRulesExecutor = resourceDataRulesExecutor,
           fhirPathDataExtractor = fhirPathDataExtractor,
@@ -325,6 +331,7 @@ class QuestionnaireViewModelTest : RobolectricTest() {
         subject = capture(subjectSlot),
         bundle = capture(bundleSlot),
         questionnaire = questionnaire,
+        questionnaireConfig = questionnaireConfig,
       )
 
       fhirCarePlanGenerator.conditionallyUpdateResourceStatus(
@@ -895,6 +902,41 @@ class QuestionnaireViewModelTest : RobolectricTest() {
   }
 
   @Test
+  fun testAddPractitionerInfoAppendedCorrectlyOnEncounterResource() {
+    val encounter = Encounter().apply { this.id = "123456" }
+    encounter.appendPractitionerInfo("12345")
+    Assert.assertEquals("Practitioner/12345", encounter.participant.first().individual.reference)
+  }
+
+  @Test
+  fun testAddPractitionerInfoAppendedCorrectlyOnObservationResource() {
+    val observation = Observation().apply { this.id = "123456" }
+    observation.appendPractitionerInfo("12345")
+    Assert.assertEquals("Practitioner/12345", observation.performer.first().reference)
+  }
+
+  @Test
+  fun testAddPractitionerInfoAppendedCorrectlyOnQuestionnaireResponse() {
+    val questionnaireResponse = QuestionnaireResponse().apply { this.id = "123456" }
+    questionnaireResponse.appendPractitionerInfo("12345")
+    Assert.assertEquals("Practitioner/12345", questionnaireResponse.author.reference)
+  }
+
+  @Test
+  fun testAddPractitionerInfoAppendedCorrectlyOnPatientResource() {
+    val patient = Patient().apply { Patient@ this.id = "123456" }
+    patient.appendPractitionerInfo("12345")
+    Assert.assertEquals("Practitioner/12345", patient.generalPractitioner.first().reference)
+  }
+
+  @Test
+  fun testAddPractitionerInfoAppendedCorrectlyOnFlag() {
+    val flag = Flag().apply { this.id = "123456" }
+    flag.appendPractitionerInfo("12345")
+    Assert.assertEquals("Practitioner/12345", flag.author.reference)
+  }
+
+  @Test
   fun testSaveExtractedResourcesForEditedQuestionnaire() = runTest {
     val questionnaire = extractionQuestionnaire()
     val questionnaireResponse =
@@ -993,5 +1035,53 @@ class QuestionnaireViewModelTest : RobolectricTest() {
     coVerify(exactly = 0) {
       defaultRepository.addOrUpdate(addMandatoryTags = true, resource = questionnaireResponse)
     }
+  }
+
+  @Test
+  fun testLoadCqlInputResourcesFromQuestionnaireConfig() = runBlocking {
+    val bundle = Bundle()
+
+    // Define the expected CQL input resources
+    val expectedCqlInputResources = listOf("basic-resource-id")
+
+    val questionnaireConfigCqlInputResources =
+      questionnaireConfig.copy(cqlInputResources = listOf("basic-resource-id"))
+
+    // Create a sample questionnaire with a CQL library extension
+    val questionnaire =
+      samplePatientRegisterQuestionnaire.copy().apply {
+        addExtension(
+          Extension().apply {
+            url = "https://sample.cqf-library.url"
+            setValue(StringType("Library/123"))
+          },
+        )
+      }
+
+    // Mock the retrieval of a Basic resource with the specified ID
+    val resource1 = Faker.buildBasicResource("basic-resource-id")
+    coEvery { fhirEngine.get<Basic>(any()) } answers { resource1 }
+
+    // Load the CQL input resources from the questionnaireConfig
+    val loadedCqlInputResources = questionnaireConfigCqlInputResources.cqlInputResources
+
+    // Verify that the loadedCqlInputResources match the expected list
+    Assert.assertEquals(expectedCqlInputResources, loadedCqlInputResources)
+
+    // Execute CQL by invoking the questionnaireViewModel.executeCql method
+    questionnaireViewModel.executeCql(
+      patient,
+      bundle,
+      questionnaire,
+      questionnaireConfigCqlInputResources,
+    )
+
+    // Verify that the bundle contains the expected Basic resource with ID "basic-resource-id"
+    Assert.assertTrue(
+      bundle.entry.any { it.resource is Basic && it.resource.id == "basic-resource-id" },
+    )
+
+    // Verify that the libraryEvaluator.runCqlLibrary was called with the correct arguments
+    coVerify { libraryEvaluator.runCqlLibrary("123", patient, bundle) }
   }
 }