+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/register/RegisterFragmentTest.kt
@@ -34,6 +34,7 @@ import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
 import io.mockk.runs
+import io.mockk.slot
 import io.mockk.spyk
 import io.mockk.verify
 import javax.inject.Inject
@@ -242,6 +243,26 @@ class RegisterFragmentTest : RobolectricTest() {
     runBlocking { registerFragmentSpy.handleQuestionnaireSubmission(questionnaireSubmission) }
     coVerify { registerFragmentSpy.refreshRegisterData() }
     coVerify { registerViewModel.emitSnackBarState(snackBarMessageConfig) }
+  }
+
+  @Test
+  fun testHandleQuestionnaireSubmissionUpdatesFilterResourceConfig() {
+    val questionnaireConfig =
+      QuestionnaireConfig(id = "add-member", saveQuestionnaireResponse = false)
+    val questionnaireResponse = QuestionnaireResponse().apply { id = "1234" }
+    val questionnaireSubmission =
+      QuestionnaireSubmission(
+        questionnaireConfig = questionnaireConfig,
+        questionnaireResponse = questionnaireResponse,
+      )
+    val registerFragmentSpy = spyk(registerFragment)
+
+    runBlocking { registerFragmentSpy.handleQuestionnaireSubmission(questionnaireSubmission) }
+
+    // Refresh data is called with QuestionnaireResponse param
+    val submissionSlot = slot<QuestionnaireResponse>()
+    coVerify { registerFragmentSpy.refreshRegisterData(capture(submissionSlot)) }
+    Assert.assertEquals(questionnaireResponse.id, submissionSlot.captured.id)
   }
 
   @Test