+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/register/RegisterViewModelTest.kt
@@ -17,6 +17,8 @@
 package org.smartregister.fhircore.quest.ui.register
 
 import androidx.compose.runtime.mutableStateOf
+import ca.uhn.fhir.rest.param.ParamPrefixEnum
+import com.google.android.fhir.search.StringFilterModifier
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import io.mockk.coEvery
@@ -28,15 +30,29 @@ import io.mockk.spyk
 import io.mockk.verify
 import javax.inject.Inject
 import kotlinx.coroutines.test.runTest
+import org.hl7.fhir.r4.model.Coding
+import org.hl7.fhir.r4.model.DateType
+import org.hl7.fhir.r4.model.Enumerations
+import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.hl7.fhir.r4.model.ResourceType
+import org.hl7.fhir.r4.model.StringType
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
+import org.smartregister.fhircore.engine.configuration.QuestionnaireConfig
 import org.smartregister.fhircore.engine.configuration.register.RegisterConfiguration
+import org.smartregister.fhircore.engine.configuration.register.RegisterFilterConfig
+import org.smartregister.fhircore.engine.configuration.register.RegisterFilterField
+import org.smartregister.fhircore.engine.configuration.workflow.ActionTrigger
+import org.smartregister.fhircore.engine.configuration.workflow.ApplicationWorkflow
 import org.smartregister.fhircore.engine.data.local.register.RegisterRepository
+import org.smartregister.fhircore.engine.domain.model.ActionConfig
+import org.smartregister.fhircore.engine.domain.model.Code
+import org.smartregister.fhircore.engine.domain.model.DataQuery
 import org.smartregister.fhircore.engine.domain.model.FhirResourceConfig
+import org.smartregister.fhircore.engine.domain.model.FilterCriterionConfig
 import org.smartregister.fhircore.engine.domain.model.ResourceConfig
 import org.smartregister.fhircore.engine.rulesengine.ResourceDataRulesExecutor
 import org.smartregister.fhircore.engine.util.SharedPreferenceKey
@@ -73,6 +89,13 @@ class RegisterViewModelTest : RobolectricTest() {
         ),
       )
 
+    every {
+      sharedPreferencesHelper.read(SharedPreferenceKey.LAST_SYNC_TIMESTAMP.name, null)
+    } returns "Mar 20, 03:01PM"
+  }
+
+  @Test
+  fun testPaginateRegisterData() {
     every { registerViewModel.retrieveRegisterConfiguration(any()) } returns
       RegisterConfiguration(
         appId = "app",
@@ -81,13 +104,6 @@ class RegisterViewModelTest : RobolectricTest() {
           FhirResourceConfig(baseResource = ResourceConfig(resource = ResourceType.Patient)),
         pageSize = 10,
       )
-    every {
-      sharedPreferencesHelper.read(SharedPreferenceKey.LAST_SYNC_TIMESTAMP.name, null)
-    } returns "Mar 20, 03:01PM"
-  }
-
-  @Test
-  fun testPaginateRegisterData() {
     registerViewModel.paginateRegisterData(registerId, false)
     val paginatedRegisterData = registerViewModel.paginatedRegisterData.value
     Assert.assertNotNull(paginatedRegisterData)
@@ -97,6 +113,14 @@ class RegisterViewModelTest : RobolectricTest() {
   @Test
   @kotlinx.coroutines.ExperimentalCoroutinesApi
   fun testRetrieveRegisterUiState() = runTest {
+    every { registerViewModel.retrieveRegisterConfiguration(any()) } returns
+      RegisterConfiguration(
+        appId = "app",
+        id = registerId,
+        fhirResource =
+          FhirResourceConfig(baseResource = ResourceConfig(resource = ResourceType.Patient)),
+        pageSize = 10,
+      )
     every { registerViewModel.paginateRegisterData(any(), any()) } just runs
     coEvery { registerRepository.countRegisterData(any()) } returns 200
     registerViewModel.retrieveRegisterUiState(
@@ -119,6 +143,14 @@ class RegisterViewModelTest : RobolectricTest() {
 
   @Test
   fun testOnEventSearchRegister() {
+    every { registerViewModel.retrieveRegisterConfiguration(any()) } returns
+      RegisterConfiguration(
+        appId = "app",
+        id = registerId,
+        fhirResource =
+          FhirResourceConfig(baseResource = ResourceConfig(resource = ResourceType.Patient)),
+        pageSize = 10,
+      )
     every { registerViewModel.registerUiState } returns
       mutableStateOf(RegisterUiState(registerId = registerId))
     // Search with empty string should paginate the data
@@ -150,5 +182,229 @@ class RegisterViewModelTest : RobolectricTest() {
     registerViewModel.onEvent(RegisterEvent.MoveToPreviousPage)
     Assert.assertEquals(1, registerViewModel.currentPage.value)
     verify { registerViewModel.paginateRegisterData(any(), any()) }
+  }
+
+  // TODO Test (CHT) remaining scenarios (e.g. filter by quantity, reference, number, datetime etc)
+  @Test
+  fun testUpdateRegisterFilterStateShouldUpdateFilterState() {
+    // Get state before the resource config is updated
+    val registerFilterStateBefore = registerViewModel.registerFilterState.value
+
+    val questionnaireConfig =
+      QuestionnaireConfig(id = "register-patient", saveQuestionnaireResponse = false)
+
+    val fhirResourceConfig = filterFhirResourceConfig()
+
+    // QuestionnaireResponse linkId should match the register filter field IDs
+    val questionnaireResponse =
+      QuestionnaireResponse().apply {
+        id = "1234"
+        addItem(
+          QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
+            linkId = "gender-filter"
+            addAnswer(
+              QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
+                value = StringType("female")
+              },
+            )
+          },
+        )
+        addItem(
+          QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
+            linkId = "birthdate-filter"
+            addAnswer(
+              QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
+                value = DateType("2000-01-01")
+              },
+            )
+          },
+        )
+        addItem(
+          QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
+            linkId = "task-status-filter"
+            addAnswer(
+              QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent().apply {
+                value = Coding("http://hl7.org/fhir/task-status", "ready", "Task status")
+              },
+            )
+          },
+        )
+      }
+
+    val dataFilterFields = registerFilterFields()
+    every { registerViewModel.retrieveRegisterConfiguration(registerId) } returns
+      RegisterConfiguration(
+        appId = "app",
+        id = registerId,
+        fhirResource = fhirResourceConfig,
+        pageSize = 10,
+        registerFilter =
+          RegisterFilterConfig(
+            dataFilterActions =
+              listOf(
+                ActionConfig(
+                  trigger = ActionTrigger.ON_CLICK,
+                  workflow = ApplicationWorkflow.LAUNCH_QUESTIONNAIRE.name,
+                  questionnaire = questionnaireConfig,
+                ),
+              ),
+            dataFilterFields = dataFilterFields,
+          ),
+      )
+    registerViewModel.updateRegisterFilterState(
+      registerId = registerId,
+      questionnaireResponse = questionnaireResponse,
+    )
+
+    val registerFilterStateAfter = registerViewModel.registerFilterState.value
+    Assert.assertNull(registerFilterStateBefore.fhirResourceConfig)
+    val updatedFhirResourceConfig = registerFilterStateAfter.fhirResourceConfig
+    Assert.assertNotNull(updatedFhirResourceConfig)
+
+    // Data queries assertions to confirm that the queries were updated.
+    // If no queries existed before filter then new ones will be created.
+    val newBaseResourceQueries = updatedFhirResourceConfig?.baseResource?.dataQueries
+    Assert.assertNotNull(newBaseResourceQueries)
+    Assert.assertNotNull(
+      newBaseResourceQueries!!.find { dataQuery ->
+        dataQuery.paramName == "gender" &&
+          dataQuery.filterCriteria.any {
+            it.dataType == Enumerations.DataType.STRING && it.value == "female"
+          }
+      },
+    )
+
+    Assert.assertNotNull(
+      newBaseResourceQueries.find { dataQuery ->
+        dataQuery.paramName == "birthdate" &&
+          dataQuery.filterCriteria.any {
+            it.dataType == Enumerations.DataType.DATE && it.value == "2000-01-01"
+          }
+      },
+    )
+
+    val taskRelatedResourceDataQueries =
+      updatedFhirResourceConfig.relatedResources
+        .find { it.id == ResourceType.Task.name }
+        ?.dataQueries
+    Assert.assertNotNull(taskRelatedResourceDataQueries)
+    Assert.assertNotNull(
+      taskRelatedResourceDataQueries!!.find { dataQuery ->
+        dataQuery.paramName == "status" &&
+          dataQuery.filterCriteria.any {
+            it.dataType == Enumerations.DataType.CODE && (it.value as Code).code == "ready"
+          }
+      },
+    )
+  }
+
+  private fun registerFilterFields() =
+    listOf(
+      RegisterFilterField(
+        filterId = ResourceType.Patient.name,
+        dataQueries =
+          listOf(
+            DataQuery(
+              paramName = "gender",
+              filterCriteria =
+                listOf(
+                  FilterCriterionConfig.StringFilterCriterionConfig(
+                    dataType = Enumerations.DataType.STRING,
+                    modifier = StringFilterModifier.MATCHES_EXACTLY,
+                    dataFilterLinkId = "gender-filter",
+                  ),
+                ),
+            ),
+            DataQuery(
+              paramName = "birthdate",
+              filterCriteria =
+                listOf(
+                  FilterCriterionConfig.DateFilterCriterionConfig(
+                    dataType = Enumerations.DataType.DATE,
+                    prefix = ParamPrefixEnum.LESSTHAN_OR_EQUALS,
+                    dataFilterLinkId = "birthdate-filter",
+                  ),
+                ),
+            ),
+          ),
+      ),
+      RegisterFilterField(
+        filterId = ResourceType.Task.name,
+        dataQueries =
+          listOf(
+            DataQuery(
+              paramName = "status",
+              filterCriteria =
+                listOf(
+                  FilterCriterionConfig.TokenFilterCriterionConfig(
+                    dataType = Enumerations.DataType.CODE,
+                    dataFilterLinkId = "task-status-filter",
+                  ),
+                ),
+            ),
+          ),
+      ),
+    )
+
+  private fun filterFhirResourceConfig() =
+    FhirResourceConfig(
+      baseResource =
+        ResourceConfig(
+          id = ResourceType.Patient.name,
+          resource = ResourceType.Patient,
+          filterId = ResourceType.Patient.name,
+        ),
+      relatedResources =
+        listOf(
+          ResourceConfig(
+            id = ResourceType.Task.name,
+            resource = ResourceType.Task,
+            filterId = ResourceType.Task.name,
+          ),
+        ),
+    )
+
+  @Test
+  fun testUpdateRegisterFilterStateWithNoAnswersForFilterQR() {
+    val registerFilterStateBefore = registerViewModel.registerFilterState
+    val questionnaireResponse =
+      QuestionnaireResponse().apply {
+        id = "1234"
+        addItem(
+          QuestionnaireResponse.QuestionnaireResponseItemComponent().apply {
+            linkId = "gender-filter"
+          },
+        )
+      }
+    val fhirResourceConfig = filterFhirResourceConfig()
+    val dataFilterFields = registerFilterFields()
+    val questionnaireConfig =
+      QuestionnaireConfig(id = "register-patient", saveQuestionnaireResponse = false)
+    every { registerViewModel.retrieveRegisterConfiguration(any()) } returns
+      RegisterConfiguration(
+        appId = "app",
+        id = registerId,
+        fhirResource = fhirResourceConfig,
+        pageSize = 10,
+        registerFilter =
+          RegisterFilterConfig(
+            dataFilterActions =
+              listOf(
+                ActionConfig(
+                  trigger = ActionTrigger.ON_CLICK,
+                  workflow = ApplicationWorkflow.LAUNCH_QUESTIONNAIRE.name,
+                  questionnaire = questionnaireConfig,
+                ),
+              ),
+            dataFilterFields = dataFilterFields,
+          ),
+      )
+
+    registerViewModel.updateRegisterFilterState(registerId, questionnaireResponse)
+
+    val registerFilterStateAfter = registerViewModel.registerFilterState
+
+    Assert.assertNull(registerFilterStateBefore.value.fhirResourceConfig?.baseResource?.dataQueries)
+    Assert.assertNull(registerFilterStateAfter.value.fhirResourceConfig?.baseResource?.dataQueries)
   }
 }